Diabetes-and-Heart-Disease-Prediction
Installations:
The project has been done in the jupyter-enviroment with python 3.11. 

Motivatoin:
To perform analysis on two different disease dataset to predict the chances of having a particular diseases by using different machine learning algorithm and also to check that which particular algorithm performs the best on the given dataset. This project uses SVC and decision tree to predict the output on the given dataset. The dataset has been taken from  diabetes dataset ''https://www.kaggle.com/code/melikedilekci/diabetes-dataset-for-beginners/input'' and heart disease  "https://www.kaggle.com/datasets/johnsmith88/heart-disease-dataset".

Summary:
Found that SVM and the Random forest algorithms. It gives an accuracy of 75% on diabetic data set and above 80% on heart disease dataset.